commit usecases

+++ backend/src/todos/use-cases/delete-todo.use-case.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from "@nestjs/common";
+import { Injectable, Logger, NotFoundException } from "@nestjs/common";
 import { DeleteTodoRepository } from "../repository";
 
 @Injectable()
@@ -9,11 +9,17 @@ export class DeleteTodoUseCases {
     ) {}
     async execute(id: string) {
         try {
-            this.logger.log('Deleting todo...');
+            this.logger.log('Deleting toDo...');
+            const todo = await this.deleteTodoRepository.findById(id);
+
+            if (!todo) {
+                throw new NotFoundException("ToDo not found");
+            }
             await this.deleteTodoRepository.delete(id);
-            this.logger.log('Todo deleted successfully.');
+            this.logger.log('ToDo deleted successfully.');
+            return todo;
         } catch (error) {
-            this.logger.error('Error deleting todo.', error);
+            this.logger.error(error);
             throw error;
         }
     }
